requirements.txt
python-dotenv
groq
elevenlabs
requests
